Item 17826 - Support optionality of class file parameter in GenePattern integration, reviewed by Eric.

+++ software/caintegrator2-war/src/gov/nih/nci/caintegrator2/application/analysis/GenePatternHelper.java
@@ -327,11 +327,13 @@ class GenePatternHelper {
 
     private void handleSampleClassification(SampleClassificationParameterValue parameterValue, 
             ParameterInfo genePatternParameter) {
-        File clsFile = 
-            new File(System.getProperty("java.io.tmpdir"), "caintegrator2_job" + tempFileCounter++ + ".cls");
-        clsFile.deleteOnExit();
-        ClassificationsToClsConverter.writeAsCls(parameterValue, clsFile.getAbsolutePath());
-        genePatternParameter.setValue(clsFile.getAbsolutePath());
+        if (!parameterValue.getClassifications().isEmpty()) {
+            File clsFile = 
+                new File(System.getProperty("java.io.tmpdir"), "caintegrator2_job" + tempFileCounter++ + ".cls");
+            clsFile.deleteOnExit();
+            ClassificationsToClsConverter.writeAsCls(parameterValue, clsFile.getAbsolutePath());
+            genePatternParameter.setValue(clsFile.getAbsolutePath());
+        }
     }
 
     private void handleGenomicData(GenomicDataParameterValue parameterValue, ParameterInfo genePatternParameter) {
